# openapi.json
{"openapi":"3.1.0","info":{"title":"rust-svelte-todo","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/auth/me":{"get":{"tags":["Auth"],"summary":"Get current user","description":"Returns the profile of the currently authenticated user.\n\nAuthentication is accepted via `auth_token` cookie (preferred) or\n`Authorization: Bearer <token>` header.","operationId":"me","responses":{"200":{"description":"Current user profile","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/auth/signin":{"post":{"tags":["Auth"],"summary":"Sign in","description":"Authenticates an existing user and sets an `auth_token` HttpOnly cookie.\n\nThe cookie is used automatically for subsequent requests. The token is also returned\nin the response body for clients that prefer the `Authorization: Bearer` header.","operationId":"signin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninUser"}}},"required":true},"responses":{"200":{"description":"Authenticated, returns JWT token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"401":{"description":"Invalid email or password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/auth/signout":{"post":{"tags":["Auth"],"summary":"Sign out","description":"Clears the `auth_token` cookie.","operationId":"signout","responses":{"204":{"description":"Signed out, cookie cleared"}}}},"/tenants":{"get":{"tags":["Tenants"],"summary":"List all tenants","description":"Returns all tenant documents from `mt_admin.tenants`.\nEach tenant describes a club or organisation hosted on the platform,\ntogether with its enabled feature modules and the name of its own database.","operationId":"list_tenants","parameters":[{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of tenants","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TenantResponse"}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/tenants/{tenant_id}":{"get":{"tags":["Tenants"],"summary":"Get tenant by ID","description":"Returns the tenant document for the specified tenant ID.","operationId":"get_tenant","parameters":[{"name":"tenant_id","in":"path","description":"Tenant ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Tenant details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Tenant not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/tenants/{tenant_id}/stats":{"get":{"tags":["Tenants"],"summary":"Stats for a tenant","description":"Returns various statistics about the tenant, such as number of users, todos, etc. (not implemented yet)","operationId":"get_tenant_stats","parameters":[{"name":"tenant_id","in":"path","description":"Tenant ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Tenant statistics","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantStatsResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Tenant not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/users":{"get":{"tags":["Users"],"summary":"List all users","description":"Returns all users in the database.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"list_users","parameters":[{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"List of users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Users"],"summary":"Create a new user","description":"Creates a new user with a unique slug generated from the full name.\n\nIf a slug collision occurs, it will attempt to generate a new slug up to 3 times\nbefore returning a conflict error.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"create_user","parameters":[{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUser"}}},"required":true},"responses":{"201":{"description":"User created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Slug already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/users/{id}":{"get":{"tags":["Users"],"summary":"Get a user by ID","description":"Returns a single user identified by its ID.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"get_user","parameters":[{"name":"id","in":"path","description":"The user ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Users"],"summary":"Update a user","description":"Updates the full name, email, and role of an existing user.\nThe slug is regenerated from the new full name. If a slug collision occurs, it will\nattempt to generate a new slug up to 3 times before returning a conflict error.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"update_user","parameters":[{"name":"id","in":"path","description":"The user ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUser"}}},"required":true},"responses":{"200":{"description":"User updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Slug already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Users"],"summary":"Delete user","description":"Deletes a user identified by its ID. Returns 204 No Content on success.\n\nThe request must include a valid Bearer token in the Authorization header for authentication\n(use the `/auth/signup` or `/auth/signin` endpoint to obtain a token).","operationId":"delete_user","parameters":[{"name":"id","in":"path","description":"The user ID","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"Bearer access token. Format: `Bearer <token>`","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"User deleted"},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"AutomationMoveBusinessCase":{"type":"object","required":["value"],"properties":{"label":{"type":["string","null"]},"value":{"type":"object"}}},"BiUrl":{"type":"object","properties":{"default":{"type":["boolean","null"]},"favorite":{"type":["boolean","null"]},"id":{"type":["string","null"]},"name":{"type":["string","null"]},"public":{"type":["boolean","null"]},"tenant":{"type":["boolean","null"]},"users":{"type":["array","null"],"items":{"type":"string"}}}},"CreateUser":{"type":"object","required":["full_name","email","password","role"],"properties":{"email":{"type":"string"},"full_name":{"type":"string"},"password":{"type":"string"},"role":{"$ref":"#/components/schemas/UserRole"}}},"DistributionEntry":{"type":"object","description":"A single slice in a percentage breakdown.","required":["label","count","percentage"],"properties":{"count":{"type":"integer","format":"int64","description":"Raw document count for this slice.","minimum":0},"label":{"type":"string","description":"Human-readable label (e.g. \"Samsung\", \"Female\", \"Prague\")."},"percentage":{"type":"number","format":"double","description":"Share of the total, 0–100 rounded to two decimal places."}}},"ErrorResponse":{"type":"object","required":["error"],"properties":{"error":{"type":"string"}}},"FansDataPoint":{"type":"object","description":"One point in the fans-over-time series.","required":["timestamp","count"],"properties":{"count":{"type":"integer","format":"int64","description":"Cumulative fan count up to and including this month.","minimum":0},"timestamp":{"type":"integer","format":"int64","description":"Unix timestamp in milliseconds (UTC midnight, 1st of the month)."}}},"FansDistributions":{"type":"object","description":"Fan breakdowns by device, city and gender.","required":["devices","cities","genders"],"properties":{"cities":{"type":"array","items":{"$ref":"#/components/schemas/DistributionEntry"},"description":"Most-common cities first; empty-string / null → \"Unknown\"."},"devices":{"type":"array","items":{"$ref":"#/components/schemas/DistributionEntry"},"description":"Most-common devices first; empty-string / null → \"Unknown\"."},"genders":{"type":"array","items":{"$ref":"#/components/schemas/DistributionEntry"},"description":"Female / Male / Unknown derived from the boolean `gender` field."}}},"FrequencyCap":{"type":"object","properties":{"enabled":{"type":["boolean","null"]},"maxEmails":{"type":["integer","null"],"format":"int32"},"periodDays":{"type":["integer","null"],"format":"int32"}}},"InternalFeatures":{"type":"object","properties":{"priceListImport":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/PriceListImport"}]}}},"OnOff":{"type":"string","enum":["ON","OFF"]},"PriceListImport":{"type":"object","properties":{"allowedUsers":{"type":["array","null"],"items":{"type":"string"}},"enabled":{"type":["boolean","null"]}}},"Security":{"type":"object","properties":{"firstLoginPasswordChange":{"type":["boolean","null"]},"twoFARequired":{"type":["boolean","null"]}}},"SigninUser":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string"},"password":{"type":"string"}}},"TenantResponse":{"type":"object","description":"Public representation of a tenant returned by the API.","required":["id","name","version","hostname","db_name","business_module","fans_module","gdpr_translation"],"properties":{"access_api_eshop_token":{"type":["string","null"]},"access_api_internal":{"type":["string","null"]},"access_api_mobile_token":{"type":["string","null"]},"access_api_oneid":{"type":["string","null"]},"access_api_ticketing":{"type":["string","null"]},"active":{"type":["boolean","null"]},"address":{"type":["string","null"]},"ai_segments":{"type":["boolean","null"]},"allow_default_gdpr":{"type":["boolean","null"]},"api_endpoint":{"type":["string","null"]},"approval_email":{"type":["string","null"]},"association_crm":{"type":["boolean","null"]},"automation_business_case_state":{"type":["boolean","null"]},"automation_move_business_case":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/AutomationMoveBusinessCase"}]},"aws_access_key":{"type":["string","null"]},"aws_address":{"type":["string","null"]},"aws_bucket_name":{"type":["string","null"]},"aws_bucket_region":{"type":["string","null"]},"aws_secret_key":{"type":["string","null"]},"bi_url":{"type":["array","null"],"items":{"$ref":"#/components/schemas/BiUrl"}},"bonus_guideline":{"type":["boolean","null"]},"business_module":{"$ref":"#/components/schemas/OnOff"},"clickhouse_password":{"type":["string","null"]},"clickhouse_username":{"type":["string","null"]},"club_logo":{"type":["string","null"]},"company":{"type":["string","null"]},"content_type":{"type":["string","null"]},"contract_email":{"type":["string","null"]},"contract_rs_template":{"type":["string","null"]},"create_fan_from_ticket_attribute":{"type":["boolean","null"]},"db_name":{"type":"string"},"default_language":{"type":["string","null"]},"dic":{"type":["string","null"]},"domain_name":{"type":["string","null"]},"email_frequency_cap":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/FrequencyCap"}]},"email_host":{"type":["string","null"]},"email_password":{"type":["string","null"]},"email_port":{"type":["integer","null"],"format":"int32"},"email_username":{"type":["string","null"]},"enigoo_api":{"type":["string","null"]},"enigoo_channel":{"type":["string","null"]},"enigoo_integration":{"type":["boolean","null"]},"enigoo_token":{"type":["string","null"]},"enigoo_token_api":{"type":["string","null"]},"esports_games_api":{"type":["string","null"]},"esports_games_key":{"type":["string","null"]},"fan_import_matching_strategy":{"type":["string","null"]},"fans_module":{"$ref":"#/components/schemas/OnOff"},"futured_integration":{"type":["boolean","null"]},"gdpr_translation":{"type":"object","description":"Deeply nested, variable-shape GDPR translation map."},"google":{"type":["boolean","null"]},"google_callback":{"type":["string","null"]},"google_callback_fe_fail":{"type":["string","null"]},"google_client_id":{"type":["string","null"]},"google_client_secret":{"type":["string","null"]},"highlight_before_expiration":{"type":["integer","null"],"format":"int32"},"hostname":{"type":"string"},"ico":{"type":["string","null"]},"id":{"type":"string","description":"MongoDB ObjectId serialised as a 24-character hex string."},"internal_features":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/InternalFeatures"}]},"invoice_notification":{"type":["boolean","null"]},"invoice_notification_days_before":{"type":["integer","null"],"format":"int32"},"invoice_notification_email":{"type":["string","null"]},"is_gina_enabled":{"type":["boolean","null"]},"max_segment_recalculate":{"type":["integer","null"],"format":"int32"},"message_frequency_cap":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/FrequencyCap"}]},"message_provider":{"type":["string","null"]},"microsoft":{"type":["boolean","null"]},"microsoft_callback":{"type":["string","null"]},"microsoft_callback_fe_fail":{"type":["string","null"]},"microsoft_client_id":{"type":["string","null"]},"microsoft_client_secret":{"type":["string","null"]},"microsoft_tenant_id":{"type":["string","null"]},"mobile_module":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/OnOff"}]},"name":{"type":"string"},"neon_integration":{"type":["boolean","null"]},"nh_sms_gateway_key":{"type":["string","null"]},"one_signal_api_key":{"type":["string","null"]},"one_signal_app_id":{"type":["string","null"]},"oneid_api":{"type":["string","null"]},"oneid_integration":{"type":["boolean","null"]},"oneid_pass":{"type":["string","null"]},"oneid_token":{"type":["string","null"]},"oneid_user":{"type":["string","null"]},"only_owner_can_edit_business_case":{"type":["boolean","null"]},"plan":{"type":["string","null"]},"push_notification_frequency_cap":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/FrequencyCap"}]},"rabbitmq_consumer_count":{"type":["integer","null"],"format":"int32"},"rabbitmq_ip":{"type":["string","null"]},"rabbitmq_multiple_prefetch":{"type":["integer","null"],"format":"int32"},"rabbitmq_pass":{"type":["string","null"]},"rabbitmq_single_prefetch":{"type":["integer","null"],"format":"int32"},"rabbitmq_user":{"type":["string","null"]},"rabbitmq_vhost":{"type":["string","null"]},"security":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/Security"}]},"spark_post_api_key":{"type":["string","null"]},"spark_post_sub_account":{"type":["string","null"]},"spark_post_webhook_pass":{"type":["string","null"]},"spark_post_webhook_username":{"type":["string","null"]},"spartaid_integration":{"type":["boolean","null"]},"subject_prefixes":{"type":["array","null"],"items":{"type":"object"}},"ticket_portal_integration":{"type":["boolean","null"]},"version":{"type":"integer","format":"int32"},"visualization":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/Visualization"}]},"webhooks":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/Webhooks"}]}}},"TenantStatsResponse":{"type":"object","required":["fans_count","new_fans_last_month","fans_over_time","distributions"],"properties":{"distributions":{"$ref":"#/components/schemas/FansDistributions","description":"Breakdowns by device, city, and gender."},"fans_count":{"type":"integer","format":"int64","description":"Total number of fans in the tenant database.","minimum":0},"fans_over_time":{"type":"array","items":{"$ref":"#/components/schemas/FansDataPoint"},"description":"Cumulative monthly fan counts, one point per month (1st of month, UTC).\nStarts one month before the first registration (count = 0)."},"new_fans_last_month":{"type":"integer","format":"int64","description":"Number of new fans registered in the last 30 days.","minimum":0}}},"Token":{"type":"object","required":["token"],"properties":{"token":{"type":"string"}}},"UpdateUser":{"type":"object","required":["full_name","role"],"properties":{"full_name":{"type":"string"},"role":{"$ref":"#/components/schemas/UserRole"}}},"User":{"type":"object","required":["id","slug","full_name","email","role"],"properties":{"email":{"type":"string"},"full_name":{"type":"string"},"id":{"type":"string"},"role":{"$ref":"#/components/schemas/UserRole"},"slug":{"type":"string"}}},"UserRole":{"type":"string","enum":["admin","editor","viewer"]},"Visualization":{"type":"object","properties":{"CARD":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/VisualizationItem"}]},"PLAYER":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/VisualizationItem"}]},"STADIUM":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/VisualizationItem"}]},"STADIUM_2":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/VisualizationItem"}]}}},"VisualizationItem":{"type":"object","properties":{"svgHeight":{"type":["integer","null"],"format":"int32","minimum":0},"svgWidth":{"type":["integer","null"],"format":"int32","minimum":0},"url":{"type":["string","null"]}}},"Webhooks":{"type":"object","properties":{"ESHOP_ORDER_PAID":{"type":["string","null"]},"EVENT_CREATE":{"type":["string","null"]},"GDPR":{"type":["string","null"]},"LOYALTY_POINTS_ADDED":{"type":["string","null"]},"MEMBERSHIP_ORDER_PAID":{"type":["string","null"]},"MEMBERSHIP_STATUS_CHANGE":{"type":["string","null"]},"MEMBERSHIP_STORNO":{"type":["string","null"]},"MEMBERSHIP_SUBSCRIPTION_PAYMENT":{"type":["string","null"]},"MEMBERSHIP_SUBSCRIPTION_PAYMENT_CHANGE":{"type":["string","null"]},"MOBILE_APP_PAIR_SEASON_TICKET":{"type":["string","null"]},"MOBILE_APP_UNPAIR_SEASON_TICKET":{"type":["string","null"]},"SEASON_BARCODE_CHANGE":{"type":["string","null"]},"SEASON_ORDER_PAID":{"type":["string","null"]},"SEASON_STORNO":{"type":["string","null"]},"SEASON_TICKET_GIFT":{"type":["string","null"]},"TICKET_BARCODE_CHANGE":{"type":["string","null"]},"TICKET_ORDER_PAID":{"type":["string","null"]},"TICKET_STORNO":{"type":["string","null"]},"TURNSTILE_ENTRANCE":{"type":["string","null"]},"USER_ONEID_DELETE":{"type":["string","null"]},"USER_ONEID_REGISTRATION":{"type":["string","null"]}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[{"name":"Auth"},{"name":"Users"},{"name":"Tenants","description":"Multi-tenant registry and per-tenant fan counts (Atlas / MongoDB)"}]}